+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/directconnectivity/GatewayAddressCache.java
@@ -987,10 +987,9 @@ public class GatewayAddressCache implements IAddressCache {
 
                 // replica validation should be triggered for an address with Unknown health status only when
                 // the address is used by a container / collection which was part of the warm up flow
-                if (addressToBeValidated.getHealthStatus() == Uri.HealthStatus.Unknown) {
-                    if (!this.proactiveOpenConnectionsProcessor.isCollectionRidUnderOpenConnectionsFlow(collectionRid)) {
+                if (addressToBeValidated.getHealthStatus() == Uri.HealthStatus.Unknown
+                    && !this.proactiveOpenConnectionsProcessor.isCollectionRidUnderOpenConnectionsFlow(collectionRid)) {
                         continue;
-                    }
                 }
 
                 Mono.fromFuture(this.proactiveOpenConnectionsProcessor